fix: MSW api not finding relevant timestamp (within last 3hrs) fixed
When parsing data from magic seaweed, the function looks for any
timestamps from within the last 3 hours and renders the data from that
particular object. When there is a gap between data which is greater than
3 hours or there is no past data, the function finds no data and therefore
cannot render the info. The fix compensates for this by parsing the nearest
future data (within 3hrs).