fix: Change error content/length
This was changed because when viewport window size was increased, the error window would shrink leaving a space between it and the beginning of the location grid. Instead of adding a fix to this, it was decided to make all of the  error message approx. the same size to avoid having to create multiple
starting points for the location grid depending on the error msg size.