refactor: Remove true/false variables -->
These particular variables would indicate when an api request was processing,
the surf conditionsguide was open, or if set timeout was still counting down.
These were to ensure multiple requests weren't bombarding the DOM with delayed
rendering of new objects. Now that buttons have been installed, this isn't a
problem. Though, it can still happen until clicking of other locations is
disabled when viewing the surf guide and subsequently the surf conditions.