fix: Change when map bounds are set
When hiding and showing the map, the map bounds were reset even on a hidden
map. For some reason when the map would be eventually displayed again, the map would often be skewed left and the info-windows would be very small (too small
to read the text they held). In order to fix this, conditional statements were
placed in the map hide/display button and at the end of any search requests
specifying to only reset bounds when the map is displayed.