fix: Set 'favorite' icon as a hard coded inline svg in index.html
In the previous configuration, the 'favorite' icon was sometimes converted to
an svg too late or after the update location frame with favorites function was
run, which would result in a white, colorless 'favorite' icon on each location
frame. Attempts to wait until the icon was converted to an inline svg before
updating the frames also failed stylistically. It resulted in split second
flashes of unstyled 'favorite' icons. Hard coding it into the index.html page
seems to fix this problemm best.